+++ src/core_modules/capture-core/components/Pages/common/TEIRelationshipsWidget/RegisterTei/DataEntry/TrackedEntityInstance/index.js
@@ -1,2 +1,3 @@
 // @flow
 export { DataEntryTrackedEntityInstance } from './DataEntryTrackedEntityInstance';
+export type { TeiPayload } from './dataEntryTrackedEntityInstance.types';