+++ src/core_modules/capture-core/components/ScopeSelector/QuickSelector/QuickSelector.types.js
@@ -17,4 +17,5 @@ export type Props = {
     onStartAgain: () => void,
     formIsOpen: boolean,
     children: Node,
+    isReadOnly?: boolean,
 };