+++ src/core_modules/capture-core/components/WidgetRelatedStages/EnterDataInOrgUnit/EnterData.component.js
@@ -48,7 +48,6 @@ type Props = {
     linkableStageLabel: string,
     relatedStagesDataValues: RelatedStageDataValueStates,
     setRelatedStagesDataValues: (() => Object) => void,
-    currentStageLabel: string,
     saveAttempted: boolean,
     errorMessages: ErrorMessagesForRelatedStages,
     ...CssClasses
@@ -60,7 +59,6 @@ export const EnterDataInOrgUnitPlain = ({
     setRelatedStagesDataValues,
     saveAttempted,
     errorMessages,
-    currentStageLabel,
     classes,
 }: Props) => {
     const onSelectOrgUnit = (e: { id: string, displayName: string, path: string }) => {
@@ -98,11 +96,10 @@ export const EnterDataInOrgUnitPlain = ({
                 <IconInfo16 />
                 {i18n.t(
                     relatedStagesDataValues?.orgUnit?.name
-                        ? 'Enter {{linkableStageLabel}} details for {{orgUnitLabel}} in the next step after completing this {{currentStageLabel}}.'
-                        : 'Select organisation unit and enter {{linkableStageLabel}} details in the next step after completing this {{currentStageLabel}}.',
+                        ? 'Enter {{linkableStageLabel}} details for {{orgUnitLabel}} in the next step'
+                        : 'Select organisation unit and enter {{linkableStageLabel}} details in the next step',
                     {
                         linkableStageLabel,
-                        currentStageLabel,
                         orgUnitLabel: relatedStagesDataValues?.orgUnit?.name,
                     },
                 )}